# docs/api/openapi-legalpro-node.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LegalPro Node API",
    "version": "1.0.0",
    "description": "API backend Node.js de LegalPro - auth, organizaciones, IA, minimax, ARCO, datos personales",
    "contact": { "name": "LegalPro", "email": "api@legalpro.pe" }
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Dev local" },
    { "url": "https://legalpro-node.railway.app", "description": "Produccion" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/expedientes": {
      "get": { "summary": "Listar expedientes (paginado, multi-tenant)", "tags": ["expedientes"], "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }, { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }, { "name": "materia", "in": "query", "schema": { "type": "string" } }, { "name": "search", "in": "query", "schema": { "type": "string" } }], "responses": { "200": { "description": "OK" } } },
      "post": { "summary": "Crear expediente", "tags": ["expedientes"], "responses": { "201": { "description": "Creado" } } }
    },
    "/api/expedientes/{id}": {
      "get": { "summary": "Obtener expediente por ID", "tags": ["expedientes"], "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "OK" }, "404": { "description": "No encontrado" } } }
    },
...
  }
}